+++ tencentcloud/apigateway/v20180808/errors.go
@@ -407,6 +407,9 @@ const (
 	// 密钥已绑定使用计划。
 	UNSUPPORTEDOPERATION_ALREADYBINDUSAGEPLAN = "UnsupportedOperation.AlreadyBindUsagePlan"
 
+	// 当前插件不支持绑定。
+	UNSUPPORTEDOPERATION_ATTACHPLUGIN = "UnsupportedOperation.AttachPlugin"
+
 	// 协议为HTTP时，不支持强制Https。
 	UNSUPPORTEDOPERATION_FORCEHTTPS = "UnsupportedOperation.ForceHttps"