+++ tencentcloud/essbasic/v20210526/models.go
@@ -724,7 +724,7 @@ type Recipient struct {
 
 type ResourceUrlInfo struct {
 
-	// 资源链接地址
+	// 资源链接地址，过期时间5分钟
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	Url *string `json:"Url,omitempty" name:"Url"`